# generated/python-flask/openapi_server/models/pokemon.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Pokemon(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, name=None, type=None, level=None, is_legendary=False, created_at=None, updated_at=None):  # noqa: E501
        """Pokemon - a model defined in OpenAPI

        :param id: The id of this Pokemon.  # noqa: E501
        :type id: UUID
        :param name: The name of this Pokemon.  # noqa: E501
        :type name: str
        :param type: The type of this Pokemon.  # noqa: E501
        :type type: str
        :param level: The level of this Pokemon.  # noqa: E501
        :type level: int
        :param is_legendary: The is_legendary of this Pokemon.  # noqa: E501
        :type is_legendary: bool
        :param created_at: The created_at of this Pokemon.  # noqa: E501
        :type created_at: datetime
        :param updated_at: The updated_at of this Pokemon.  # noqa: E501
        :type updated_at: datetime
        """
        self.openapi_types = {
            'id': UUID,
            'name': str,
            'type': str,
            'level': int,
            'is_legendary': bool,
            'created_at': datetime,
            'updated_at': datetime
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'type': 'type',
            'level': 'level',
            'is_legendary': 'isLegendary',
            'created_at': 'createdAt',
            'updated_at': 'updatedAt'
        }

        self._id = id
        self._name = name
        self._type = type
        self._level = level
        self._is_legendary = is_legendary
        self._created_at = created_at
        self._updated_at = updated_at

    @classmethod
    def from_dict(cls, dikt) -> 'Pokemon':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Pokemon of this Pokemon.  # noqa: E501
        :rtype: Pokemon
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> UUID:
        """Gets the id of this Pokemon.


        :return: The id of this Pokemon.
        :rtype: UUID
        """
        return self._id

    @id.setter
    def id(self, id: UUID):
        """Sets the id of this Pokemon.


        :param id: The id of this Pokemon.
        :type id: UUID
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Pokemon.


        :return: The name of this Pokemon.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Pokemon.


        :param name: The name of this Pokemon.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def type(self) -> str:
        """Gets the type of this Pokemon.


        :return: The type of this Pokemon.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this Pokemon.


        :param type: The type of this Pokemon.
        :type type: str
        """
        if type is None:
            raise ValueError("Invalid value for `type`, must not be `None`")  # noqa: E501

        self._type = type

    @property
    def level(self) -> int:
        """Gets the level of this Pokemon.


        :return: The level of this Pokemon.
        :rtype: int
        """
        return self._level

    @level.setter
    def level(self, level: int):
        """Sets the level of this Pokemon.


        :param level: The level of this Pokemon.
        :type level: int
        """
        if level is None:
            raise ValueError("Invalid value for `level`, must not be `None`")  # noqa: E501
        if level is not None and level > 100:  # noqa: E501
            raise ValueError("Invalid value for `level`, must be a value less than or equal to `100`")  # noqa: E501
        if level is not None and level < 1:  # noqa: E501
            raise ValueError("Invalid value for `level`, must be a value greater than or equal to `1`")  # noqa: E501

        self._level = level

    @property
    def is_legendary(self) -> bool:
        """Gets the is_legendary of this Pokemon.


        :return: The is_legendary of this Pokemon.
        :rtype: bool
        """
        return self._is_legendary

    @is_legendary.setter
    def is_legendary(self, is_legendary: bool):
        """Sets the is_legendary of this Pokemon.


        :param is_legendary: The is_legendary of this Pokemon.
        :type is_legendary: bool
        """

        self._is_legendary = is_legendary

    @property
    def created_at(self) -> datetime:
        """Gets the created_at of this Pokemon.


        :return: The created_at of this Pokemon.
        :rtype: datetime
        """
        return self._created_at

    @created_at.setter
    def created_at(self, created_at: datetime):
        """Sets the created_at of this Pokemon.


        :param created_at: The created_at of this Pokemon.
        :type created_at: datetime
        """

        self._created_at = created_at

    @property
    def updated_at(self) -> datetime:
        """Gets the updated_at of this Pokemon.


        :return: The updated_at of this Pokemon.
        :rtype: datetime
        """
        return self._updated_at

    @updated_at.setter
    def updated_at(self, updated_at: datetime):
        """Sets the updated_at of this Pokemon.


        :param updated_at: The updated_at of this Pokemon.
        :type updated_at: datetime
        """

        self._updated_at = updated_at