# generated/python-flask/openapi_server/models/pokemon_list_response.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server.models.pokemon import Pokemon
from openapi_server import util

from openapi_server.models.pokemon import Pokemon  # noqa: E501

class PokemonListResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, items=None, total=None, limit=None, offset=None):  # noqa: E501
        """PokemonListResponse - a model defined in OpenAPI

        :param items: The items of this PokemonListResponse.  # noqa: E501
        :type items: List[Pokemon]
        :param total: The total of this PokemonListResponse.  # noqa: E501
        :type total: int
        :param limit: The limit of this PokemonListResponse.  # noqa: E501
        :type limit: int
        :param offset: The offset of this PokemonListResponse.  # noqa: E501
        :type offset: int
        """
        self.openapi_types = {
            'items': List[Pokemon],
            'total': int,
            'limit': int,
            'offset': int
        }

        self.attribute_map = {
            'items': 'items',
            'total': 'total',
            'limit': 'limit',
            'offset': 'offset'
        }

        self._items = items
        self._total = total
        self._limit = limit
        self._offset = offset

    @classmethod
    def from_dict(cls, dikt) -> 'PokemonListResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PokemonListResponse of this PokemonListResponse.  # noqa: E501
        :rtype: PokemonListResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def items(self) -> List[Pokemon]:
        """Gets the items of this PokemonListResponse.


        :return: The items of this PokemonListResponse.
        :rtype: List[Pokemon]
        """
        return self._items

    @items.setter
    def items(self, items: List[Pokemon]):
        """Sets the items of this PokemonListResponse.


        :param items: The items of this PokemonListResponse.
        :type items: List[Pokemon]
        """
        if items is None:
            raise ValueError("Invalid value for `items`, must not be `None`")  # noqa: E501

        self._items = items

    @property
    def total(self) -> int:
        """Gets the total of this PokemonListResponse.


        :return: The total of this PokemonListResponse.
        :rtype: int
        """
        return self._total

    @total.setter
    def total(self, total: int):
        """Sets the total of this PokemonListResponse.


        :param total: The total of this PokemonListResponse.
        :type total: int
        """
        if total is None:
            raise ValueError("Invalid value for `total`, must not be `None`")  # noqa: E501
        if total is not None and total < 0:  # noqa: E501
            raise ValueError("Invalid value for `total`, must be a value greater than or equal to `0`")  # noqa: E501

        self._total = total

    @property
    def limit(self) -> int:
        """Gets the limit of this PokemonListResponse.


        :return: The limit of this PokemonListResponse.
        :rtype: int
        """
        return self._limit

    @limit.setter
    def limit(self, limit: int):
        """Sets the limit of this PokemonListResponse.


        :param limit: The limit of this PokemonListResponse.
        :type limit: int
        """
        if limit is None:
            raise ValueError("Invalid value for `limit`, must not be `None`")  # noqa: E501
        if limit is not None and limit < 1:  # noqa: E501
            raise ValueError("Invalid value for `limit`, must be a value greater than or equal to `1`")  # noqa: E501

        self._limit = limit

    @property
    def offset(self) -> int:
        """Gets the offset of this PokemonListResponse.


        :return: The offset of this PokemonListResponse.
        :rtype: int
        """
        return self._offset

    @offset.setter
    def offset(self, offset: int):
        """Sets the offset of this PokemonListResponse.


        :param offset: The offset of this PokemonListResponse.
        :type offset: int
        """
        if offset is None:
            raise ValueError("Invalid value for `offset`, must not be `None`")  # noqa: E501
        if offset is not None and offset < 0:  # noqa: E501
            raise ValueError("Invalid value for `offset`, must be a value greater than or equal to `0`")  # noqa: E501

        self._offset = offset