# generated/python-flask/openapi_server/models/pokemon_update_request.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class PokemonUpdateRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, type=None, level=None, is_legendary=None):  # noqa: E501
        """PokemonUpdateRequest - a model defined in OpenAPI

        :param name: The name of this PokemonUpdateRequest.  # noqa: E501
        :type name: str
        :param type: The type of this PokemonUpdateRequest.  # noqa: E501
        :type type: str
        :param level: The level of this PokemonUpdateRequest.  # noqa: E501
        :type level: int
        :param is_legendary: The is_legendary of this PokemonUpdateRequest.  # noqa: E501
        :type is_legendary: bool
        """
        self.openapi_types = {
            'name': str,
            'type': str,
            'level': int,
            'is_legendary': bool
        }

        self.attribute_map = {
            'name': 'name',
            'type': 'type',
            'level': 'level',
            'is_legendary': 'isLegendary'
        }

        self._name = name
        self._type = type
        self._level = level
        self._is_legendary = is_legendary

    @classmethod
    def from_dict(cls, dikt) -> 'PokemonUpdateRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PokemonUpdateRequest of this PokemonUpdateRequest.  # noqa: E501
        :rtype: PokemonUpdateRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this PokemonUpdateRequest.


        :return: The name of this PokemonUpdateRequest.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this PokemonUpdateRequest.


        :param name: The name of this PokemonUpdateRequest.
        :type name: str
        """
        if name is not None and len(name) > 50:
            raise ValueError("Invalid value for `name`, length must be less than or equal to `50`")  # noqa: E501
        if name is not None and len(name) < 1:
            raise ValueError("Invalid value for `name`, length must be greater than or equal to `1`")  # noqa: E501

        self._name = name

    @property
    def type(self) -> str:
        """Gets the type of this PokemonUpdateRequest.


        :return: The type of this PokemonUpdateRequest.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this PokemonUpdateRequest.


        :param type: The type of this PokemonUpdateRequest.
        :type type: str
        """
        if type is not None and len(type) > 50:
            raise ValueError("Invalid value for `type`, length must be less than or equal to `50`")  # noqa: E501
        if type is not None and len(type) < 1:
            raise ValueError("Invalid value for `type`, length must be greater than or equal to `1`")  # noqa: E501

        self._type = type

    @property
    def level(self) -> int:
        """Gets the level of this PokemonUpdateRequest.


        :return: The level of this PokemonUpdateRequest.
        :rtype: int
        """
        return self._level

    @level.setter
    def level(self, level: int):
        """Sets the level of this PokemonUpdateRequest.


        :param level: The level of this PokemonUpdateRequest.
        :type level: int
        """
        if level is not None and level > 100:  # noqa: E501
            raise ValueError("Invalid value for `level`, must be a value less than or equal to `100`")  # noqa: E501
        if level is not None and level < 1:  # noqa: E501
            raise ValueError("Invalid value for `level`, must be a value greater than or equal to `1`")  # noqa: E501

        self._level = level

    @property
    def is_legendary(self) -> bool:
        """Gets the is_legendary of this PokemonUpdateRequest.


        :return: The is_legendary of this PokemonUpdateRequest.
        :rtype: bool
        """
        return self._is_legendary

    @is_legendary.setter
    def is_legendary(self, is_legendary: bool):
        """Sets the is_legendary of this PokemonUpdateRequest.


        :param is_legendary: The is_legendary of this PokemonUpdateRequest.
        :type is_legendary: bool
        """

        self._is_legendary = is_legendary